requirements.txt
flask
torch
flash-attn
bitsandbytes
transformers
Pillow 
torchvision 
transformers
sentencepiece
decord 
huggingface_hub
waitress
accelerate
